Add device types configuration and endpoint
This adds a new endpoint to list all the supported device types. Device
types are configured via JSON files external to the assembly, so that
users can add custom types without rebuilding the service. The files
include a README describing the format of these files.
I was going to use HOCON for these files too, for a cleaner look, but
I'm actually reconsidering that approach and the use of Newtonsoft
greatly reduces the code required to parse these files (at least in
.NET).